requirements.txt
streamlit
sentence-transformers
numpy
scikit-learn
openai
python-dotenv
